backend/requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
docker
cryptography
Flask-Bcrypt
requests
urllib3
PyJWT